Refactor: call_core() should be a method of RaftInner instead of a method of Raft
